perf: finalize phase 1-8 auth and dashboard runtime hardening
