+++ .github/workflows/ci.yml
@@ -138,6 +138,9 @@ jobs:
       - name: Baseline Performance Snapshot
         run: npm run perf:baseline
 
+      - name: Dashboard Runtime Gate
+        run: npm run perf:dashboard-runtime
+
       - name: Lighthouse Performance Gate
         run: npm run perf:lighthouse
